+++ src/cloud/vms.ts
@@ -1,5 +1,5 @@
 import * as osc from "outscale-api";
-import { getConfig } from '../cloud/cloud';
+import { getConfig } from './cloud';
 import { Profile } from "../flat/node";
 
 export function getVms(profile: Profile): Promise<Array<osc.Vm> | string> {
